requirements.txt
affine==2.4.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.7.0
anywidget==0.9.13
APScheduler==3.11.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
bqplot==0.12.44
branca==0.8.1
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
color-operations==0.1.6
colour==0.1.5
comm==0.2.2
contextily==1.6.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
Django==5.1.4
django-apscheduler==0.7.0
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
duckdb==1.1.3
earthengine-api==1.4.3
eerepr==0.0.4
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Cors==5.0.0
flask-restx==1.3.0
folium==0.19.3
fonttools==4.55.3
future==1.0.0
gdown==5.2.0
geemap==0.35.1
geocoder==1.38.1
geographiclib==2.0
geojson==3.2.0
geopandas==1.0.1
geopy==2.4.1
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanize==4.11.0
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.5
ipyleaflet==0.19.2
ipython==8.31.0
ipytree==0.2.2
ipyvue==1.11.2
ipyvuetify==1.10.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-leaflet==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
leafmap==0.42.6
localtileserver==0.10.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mercantile==1.2.1
morecantile==6.2.0
nbformat==5.10.4
nest-asyncio==1.6.0
numexpr==2.10.2
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
psycopg2-binary==2.9.10
psygnal==0.11.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pymdown-extensions==10.13
pyogrio==0.10.0
pyparsing==3.2.1
pyperclip==1.9.0
pyproj==3.7.0
pyshp==2.3.1
PySocks==1.7.1
pystac==1.11.0
pystac-client==0.8.5
python-box==7.3.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.4.3
ratelim==0.1.6
reacton==1.9.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-click==1.8.5
rio-cogeo==5.4.1
rio-tiler==7.2.2
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
scooby==0.10.0
server-thread==0.2.0
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
solara==1.43.0
solara-server==1.43.0
solara-ui==1.43.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.1
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
whitebox==2.3.5
whiteboxgui==2.3.0
widgetsnbextension==4.0.13
xarray==2024.11.0
xyzservices==2024.9.0
